% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-PIG.r
\name{dTarget}
\alias{dTarget}
\alias{print.dTarget}
\title{Definition for S3 class \code{dTarget}}
\usage{
dTarget(priority, predictor, metag)

\method{print}{dTarget}(x, ...)
}
\arguments{
\item{priority}{a data frame}

\item{predictor}{a data frame}

\item{metag}{an 'igraph' object}

\item{x}{an object of class \code{dTarget}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{dTarget}
}
\description{
\code{dTarget} has 3 components: priority, predictor and metag.
}
\examples{
\dontrun{
dTarget(priority, predictor, metag)
}
}
\keyword{S3}
\keyword{classes}
