% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oAUCurve.r
\name{oAUCurve}
\alias{oAUCurve}
\title{Function to calculate the area under the curve}
\usage{
oAUCurve(
x,
y,
method = c("trapezoid", "step", "linear", "spline"),
subdivisions = 100L
)
}
\arguments{
\item{x}{numeric vectors giving the x-coordinates}

\item{y}{numeric vectors giving the y-coordinates}

\item{method}{the medthod used to calculate the area under curve. It
can be 'trapezoid', 'step', 'linear' (linear interpolation) and
'spline' (spline interpolation)}

\item{subdivisions}{the maximum number of subintervals}
}
\value{
a numeric value
}
\description{
\code{oAUCurve} is supposed to calculate the area under the curve.
}
\note{
none
}
\examples{
set.seed(825)
y <- abs(rnorm(1000))
x <- 1:1000/1000
oAUCurve(x,y)
}
\seealso{
\code{\link{oAUCurve}}
}
