% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oBicreate.r
\name{oBicreate}
\alias{oBicreate}
\title{Function to create a bipartitle graph}
\usage{
oBicreate(data, verbose = TRUE)
}
\arguments{
\item{data}{a data frame/matrix to create a bipartitle graph}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
It returns an igraph object.
}
\description{
\code{oBicreate} is supposed to create a bipartitle graph.
}
\note{
none
}
\examples{
data <- cbind(c(1,1,0,1), c(0,1,2,0))
ig <- oBicreate(data)
}
\seealso{
\code{\link{oBicreate}}
}
