% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oDAGinduce.r
\name{oDAGinduce}
\alias{oDAGinduce}
\title{Function to generate a DAG subgraph induced by input nodes}
\usage{
oDAGinduce(
ig,
nodes_query = NULL,
path.mode = c("all_paths", "shortest_paths", "all_shortest_paths",
"shortest_paths_epath"),
verbose = TRUE
)
}
\arguments{
\item{ig}{an object of class "igraph" to represent DAG}

\item{nodes_query}{nodes in query}

\item{path.mode}{the mode of paths induced by vertices/nodes with input
annotation data. It can be "all_paths" for all possible paths to the
root, "shortest_paths" for only one path to the root (for each node in
query), "all_shortest_paths" for all shortest paths to the root (i.e.
for each node, find all shortest paths with the equal lengths),
"shortest_paths_epath" (indeed only edges in the shortest path are
kept)}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
NULL or an induced subgraph, an object of class "igraph".
}
\description{
\code{oDAGinduce} is supposed to produce a subgraph induced by input
nodes, given a direct acyclic graph (DAG; an ontology). The input is a
graph of "igraph", nodes, and the mode defining the paths to the root
of DAG. The induced subgraph contains nodes and their ancestors along
with the defined paths to the root of DAG.
}
\note{
For the mode "shortest_paths", the induced subgraph is the most
concise, and thus informative for visualisation when there are many
nodes in query, while the mode "all_paths" results in the complete
subgraph.
}
\examples{
\dontrun{
###########################################################
ig <- oRDS('ig.GOMF', placeholder=placeholder)
ig1 <- oDAGinduce(ig, path.mode="all_paths")
ig2 <- oDAGinduce(ig, path.mode="shortest_paths")
ig3 <- oDAGinduce(ig, path.mode="all_shortest_paths")
# very slow below
ig4 <- oDAGinduce(ig, path.mode="shortest_paths_epath")
}
}
\seealso{
\code{\link{oDAGinduce}}
}
