% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oDAGroot.r
\name{oDAGroot}
\alias{oDAGroot}
\title{Function to find the root node of a direct acyclic graph (DAG)}
\usage{
oDAGroot(g)
}
\arguments{
\item{g}{an object of class "igraph" or "graphNEL"}
}
\value{
\itemize{
\item{\code{root}: the root name (i.e. Term ID)}
}
}
\description{
\code{oDAGroot} is supposed to find the root node of a direct acyclic
graph (DAG; an ontology). It return the name (i.e Term ID) of the root
node.
}
\note{
none
}
\examples{
\dontrun{
# 1) load HPPA as igraph object
ig.HPPA <-dRDataLoader(RData='ig.HPPA')
g <- ig.HPPA

# 2) find the root
root <- oDAGroot(g)
root
}
}
\seealso{
\code{\link{oDAGroot}}
}
