% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oDefineQTL.r
\name{oDefineQTL}
\alias{oDefineQTL}
\title{Function to extract eQTL-gene pairs given a list of SNPs or a customised eQTL mapping data}
\usage{
oDefineQTL(
data = NULL,
include.QTL = c(NA, "eQTL_xMEdb_Bcell", "eQTL_xMEdb_Blood",
"eQTL_xMEdb_CD14",
"eQTL_xMEdb_CD4", "eQTL_xMEdb_CD8", "eQTL_xMEdb_IFN",
"eQTL_xMEdb_LPS2",
"eQTL_xMEdb_LPS24", "eQTL_xMEdb_Monocyte", "eQTL_xMEdb_Neutrophil",
"eQTL_xMEdb_NK",
"eQTL_DICE_Bcell_naive", "eQTL_DICE_CD4_naive", "eQTL_DICE_CD4_stimu",
"eQTL_DICE_CD4_Tfh", "eQTL_DICE_CD4_Th1", "eQTL_DICE_CD4_Th17",
"eQTL_DICE_CD4_Th1h17", "eQTL_DICE_CD4_Th2", "eQTL_DICE_CD8_naive",
"eQTL_DICE_CD8_stimu", "eQTL_DICE_Macrophages_M2",
"eQTL_DICE_Monocytes",
"eQTL_DICE_NK",
"eQTL_DICE_Treg_memo", "eQTL_DICE_Treg_naive",
"eQTL_GTExV8_Adipose_Subcutaneous",
"eQTL_GTExV8_Adipose_Visceral_Omentum",
"eQTL_GTExV8_Adrenal_Gland", "eQTL_GTExV8_Artery_Aorta",
"eQTL_GTExV8_Artery_Coronary", "eQTL_GTExV8_Artery_Tibial",
"eQTL_GTExV8_Brain_Amygdala",
"eQTL_GTExV8_Brain_Anterior_cingulate_cortex_BA24",
"eQTL_GTExV8_Brain_Caudate_basal_ganglia",
"eQTL_GTExV8_Brain_Cerebellar_Hemisphere",
"eQTL_GTExV8_Brain_Cerebellum", "eQTL_GTExV8_Brain_Cortex",
"eQTL_GTExV8_Brain_Frontal_Cortex_BA9",
"eQTL_GTExV8_Brain_Hippocampus",
"eQTL_GTExV8_Brain_Hypothalamus",
"eQTL_GTExV8_Brain_Nucleus_accumbens_basal_ganglia",
"eQTL_GTExV8_Brain_Putamen_basal_ganglia",
"eQTL_GTExV8_Brain_Spinal_cord_cervical_c1",
"eQTL_GTExV8_Brain_Substantia_nigra",
"eQTL_GTExV8_Breast_Mammary_Tissue",
"eQTL_GTExV8_Cells_Cultured_fibroblasts",
"eQTL_GTExV8_Cells_EBVtransformed_lymphocytes",
"eQTL_GTExV8_Colon_Sigmoid",
"eQTL_GTExV8_Colon_Transverse",
"eQTL_GTExV8_Esophagus_Gastroesophageal_Junction",
"eQTL_GTExV8_Esophagus_Mucosa",
"eQTL_GTExV8_Esophagus_Muscularis",
"eQTL_GTExV8_Heart_Atrial_Appendage",
"eQTL_GTExV8_Heart_Left_Ventricle",
"eQTL_GTExV8_Kidney_Cortex", "eQTL_GTExV8_Liver", "eQTL_GTExV8_Lung",
"eQTL_GTExV8_Minor_Salivary_Gland", "eQTL_GTExV8_Muscle_Skeletal",
"eQTL_GTExV8_Nerve_Tibial", "eQTL_GTExV8_Ovary",
"eQTL_GTExV8_Pancreas",
"eQTL_GTExV8_Pituitary", "eQTL_GTExV8_Prostate",
"eQTL_GTExV8_Skin_Not_Sun_Exposed_Suprapubic",
"eQTL_GTExV8_Skin_Sun_Exposed_Lower_leg",
"eQTL_GTExV8_Small_Intestine_Terminal_Ileum",
"eQTL_GTExV8_Spleen",
"eQTL_GTExV8_Stomach", "eQTL_GTExV8_Testis", "eQTL_GTExV8_Thyroid",
"eQTL_GTExV8_Uterus", "eQTL_GTExV8_Vagina", "eQTL_GTExV8_Whole_Blood",
"eQTL_eQTLGen", "pQTL_Plasma", "eQTL_LCL"),
QTL.customised = NULL,
verbose = TRUE,
placeholder = NULL,
guid = NULL
)
}
\arguments{
\item{data}{NULL or an input vector containing SNPs. If NULL, all SNPs
will be considered. If a input vector containing SNPs, SNPs should be
provided as dbSNP ID (ie starting with rs). Alternatively, they can be
in the format of 'chrN:xxx', where N is either 1-22 or X, xxx is
number; for example, 'chr16:28525386'}

\item{include.QTL}{genes modulated by eQTL (also Lead SNPs or in LD
with Lead SNPs) are also included. By default, it is 'NA' to disable
this option. Otherwise, those genes modulated by eQTL will be included.
Pre-built eQTL datasets are detailed in the section 'Note'}

\item{QTL.customised}{a user-input matrix or data frame with 4 columns:
1st column for SNPs/eQTLs, 2nd column for Genes, 3rd for eQTL mapping
significance level (p-values or FDR), and 4th for contexts (required
even though only one context is input). Alternatively, it can be a file
containing these 4 columns. It is designed to allow the user analysing
their eQTL data. This customisation (if provided) will populate
built-in eQTL data; mysql -e "use pi; SELECT
rs_id_dbSNP147_GRCh37p13,gene_name,pval_nominal,Tissue FROM
GTEx_V7_pair WHERE rs_id_dbSNP147_GRCh37p13!='.';" >
/var/www/bigdata/QTL.customised.txt}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{placeholder}{the characters to tell the location of built-in RDS
files. See \code{\link{oRDS}} for details}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF
project. See \code{\link{oRDS}} for details}
}
\value{
a data frame with following columns:
\itemize{
\item{\code{SNP}: eQTLs}
\item{\code{Gene}: eQTL-containing genes}
\item{\code{Sig}: the eQTL mapping significant level}
\item{\code{Context}: the context in which eQTL data was generated}
}
}
\description{
\code{oDefineQTL} is supposed to extract eQTL-gene pairs given a list
of SNPs or a customised eQTL mapping data.
}
\note{
Pre-built eQTL datasets are described below according to the data
sources.\cr
1. eQTL_xMEdb
\itemize{
\item{\code{eQTL_xMEdb_CD14}: cis and trans-eQTLs in the resting/CD14+
state.}
\item{\code{eQTL_xMEdb_LPS2}: cis and trans-eQTLs in the activating
state induced by 2-hour LPS.}
\item{\code{eQTL_xMEdb_LPS24}: cis and trans-eQTLs in the activating
state induced by 24-hour LPS.}
\item{\code{eQTL_xMEdb_IFN}: cis and trans-eQTLs in the activating
state induced by 24-hour interferon-gamma.}
\item{\code{eQTL_xMEdb_Bcell}: cis and trans-eQTLs in B cells.}
\item{\code{eQTL_xMEdb_Blood}: cis and trans-eQTLs in the blood.}
\item{\code{eQTL_xMEdb_CD4}: cis and trans-eQTLs in the CD4 cells.}
\item{\code{eQTL_xMEdb_CD8}: cis and trans-eQTLs in the CD8 cells.}
\item{\code{eQTL_xMEdb_Monocyte}: cis and trans-eQTLs in the
monocytes.}
\item{\code{eQTL_xMEdb_Neutrophil}: cis and trans-eQTLs in the
neutrophils.}
\item{\code{eQTL_xMEdb_NK}: cis and trans-eQTLs in the NK cells.}
}
2. eQTL_DICE (PMID:30449622)
\itemize{
\item{\code{eQTL_DICE_Bcell_naive}: cis-eQTLs in the naive B cells.}
\item{\code{eQTL_DICE_CD4_naive}: cis-eQTLs in the naive CD4 cells.}
\item{\code{eQTL_DICE_CD4_stimu}: cis-eQTLs in the stimulated CD4
cells.}
\item{\code{eQTL_DICE_CD4_Tfh}: cis-eQTLs in the CD4 Tfh cells.}
\item{\code{eQTL_DICE_CD4_Th1}: cis-eQTLs in the CD4 Th1 cells.}
\item{\code{eQTL_DICE_CD4_Th17}: cis-eQTLs in the CD4 Th17 cells.}
\item{\code{eQTL_DICE_CD4_Th2}: cis-eQTLs in the CD4 Th2 cells.}
\item{\code{eQTL_DICE_CD8_naive}: cis-eQTLs in the naive CD8 cells.}
\item{\code{eQTL_DICE_CD8_stimu}: cis-eQTLs in the stimulated CD8
cells.}
\item{\code{eQTL_DICE_Macrophages_M2}: cis-eQTLs in the macrophage M2
cells.}
\item{\code{eQTL_DICE_Monocytes}: cis-eQTLs in the monocyte cells.}
\item{\code{eQTL_DICE_NK}: cis-eQTLs in the NK cells.}
\item{\code{eQTL_DICE_Treg_memo}: cis-eQTLs in the memory T regulatory
cells.}
\item{\code{eQTL_DICE_Treg_naive}: cis-eQTLs in the naive T regulatory
cells.}
}
}
\examples{
\dontrun{
# a) provide the SNPs with the significance info
gr <- ImmunoBase$AS$variants
data <- gr$Variant

# b) define eQTL genes
df_SGS <- oDefineQTL(data, include.QTL="eQTL_xMEdb_Bcell",
placeholder=placeholder)
}
}
\seealso{
\code{\link{oDefineQTL}}
}
