% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oGSEAdotplot.r
\name{oGSEAdotplot}
\alias{oGSEAdotplot}
\title{Function to visualise GSEA results using dot plot}
\usage{
oGSEAdotplot(
eGSEA,
top = 1,
colormap = "spectral",
zlim = NULL,
ncolors = 64,
xlab = NULL,
title = c("setID", "none"),
subtitle = c("leading", "enrichment", "both", "none"),
clab = "Priority\\nrating",
x.scale = c("normal", "sqrt", "log"),
peak = TRUE,
peak.color = "black",
leading = FALSE,
leading.size = 2.5,
leading.color = "steelblue4",
leading.alpha = 1,
leading.padding = 0.2,
leading.arrow = 0.01,
leading.force = 0.01,
leading.query = NULL,
leading.query.only = FALSE,
leading.edge.only = FALSE,
leading.label.direction = c("none", "left"),
compact = FALSE,
font.family = "sans",
...
)
}
\arguments{
\item{eGSEA}{an object of class "eGSEA"}

\item{top}{the number of the top enrichments to be visualised.
Alternatively, the gene set names can be queried}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{https://html-color-codes.info/color-names/index.html}}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted}

\item{ncolors}{the number of colors specified over the colormap}

\item{xlab}{the label for x-axis. If NULL, it is 'Target ranks'}

\item{title}{the title. If NULL, it is term name followed by the number
of its annotations}

\item{subtitle}{the subtitle. It can be used to show 'leading' info,
'enrichment' info or 'both'}

\item{clab}{the label for colorbar. By default, it is '5-star ratings'}

\item{x.scale}{how to transform the x scale. It can be "normal" for no
transformation, "sqrt" for square root transformation, and "log" for
log-based transformation}

\item{peak}{logical to indicate whether the peak location is shown}

\item{peak.color}{the peak color}

\item{leading}{logical to indicate whether the leading targets are
texted. Alterntively, leading can be numeric to restict the top targets
displayed}

\item{leading.size}{the size of leading targets' texts. It only works
when the parameter 'leading' is enabled}

\item{leading.color}{the label color of leading targets' texts}

\item{leading.alpha}{the 0-1 value specifying transparency of leading
targets' texts}

\item{leading.padding}{the padding around the leading targets' texts}

\item{leading.arrow}{the arrow pointing to the leading targets}

\item{leading.force}{the repelling force between leading targets'
texts}

\item{leading.query}{which genes in query will be labelled. By default,
it sets to NULL meaning all genes will be displayed. If labels in query
can not be found, then all will be displayed}

\item{leading.query.only}{logical to indicate whether only genes in
query will be displayed. By default, it sets to FALSE. It only works
when labels in query are enabled/found}

\item{leading.edge.only}{logical to indicate whether only the leading
edge will be shown. By default, it sets to FALSE}

\item{leading.label.direction}{the leading label direction. It can be
"none", "left" (aligned to the left-most edge)}

\item{compact}{logical to indicate whether the compact/void theme is
used. If TRUE, axes and legend info will be hidden}

\item{font.family}{the font family for texts}

\item{...}{additional paramters associated with
ggrepel::geom_text_repel. If queried, it has high priority (eg,
color='darkred',size=2,alpha=0.6,fontface='bold')}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE showing which function is
used to draw this graph}
}
\value{
an object of class "ggplot" or a list of "ggplot" objects.
}
\description{
\code{oGSEAdotplot} is supposed to visualise GSEA results using dot
plot. It returns an object of class "ggplot" or a list of "ggplot"
objects.
}
\note{
none
}
\examples{
\dontrun{
gp <- oGSEAdotplot(eGSEA, top=1)
#gp <- oGSEAdotplot(eGSEA, top=1, peak=FALSE, compact=TRUE, signature=FALSE)
gp

ls_gp <- oGSEAdotplot(eGSEA, top=1:4, signature=FALSE)
library(gridExtra)
grid.arrange(grobs=ls_gp, ncol=2)
}
}
\seealso{
\code{\link{oGSEAdotplot}}
}
