% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oPierGSEA.r
\name{oPierGSEA}
\alias{oPierGSEA}
\title{Function to prioritise pathways based on GSEA analysis of prioritised genes}
\usage{
oPierGSEA(
pNode,
priority.top = NULL,
customised.genesets,
size.range = c(10, 500),
type = c("simple", "multilevel"),
weight = 1,
nperm = NULL,
p.adjust.method = c("BH", "BY", "bonferroni", "holm", "hochberg",
"hommel"),
verbose = TRUE,
silent = FALSE
)
}
\arguments{
\item{pNode}{an object of class "pNode" (or "sTarget" or "dTarget").
Alternatively, it can be a data frame with two columns ('priority' and
'rank')}

\item{priority.top}{the number of the top targets used for GSEA. By
default, it is NULL meaning all targets are used}

\item{customised.genesets}{a list each containing gene symbols. By
default, it is NULL. If the list provided, it will overtake the
previous parameter "ontology"}

\item{size.range}{the minimum and maximum size of members of each term
in consideration. By default, it sets to a minimum of 10 but no more
than 500}

\item{type}{It can be 'simple' or 'multilevel'}

\item{weight}{an integer specifying score weight. It can be "0" for
unweighted (an equivalent to Kolmogorov-Smirnov, only considering the
rank), "1" for weighted by input gene score (by default), and "2" for
over-weighted, and so on}

\item{nperm}{the number of random permutations. For each permutation,
gene-score associations will be permutated so that permutation of
gene-term associations is realised}

\item{p.adjust.method}{the method used to adjust p-values. It can be
one of "BH", "BY", "bonferroni", "holm", "hochberg" and "hommel". The
first two methods "BH" (widely used) and "BY" control the false
discovery rate (FDR: the expected proportion of false discoveries
amongst the rejected hypotheses); the last four methods "bonferroni",
"holm", "hochberg" and "hommel" are designed to give strong control of
the family-wise error rate (FWER). Notes: FDR is a less stringent
condition than FWER}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true}

\item{silent}{logical to indicate whether the messages will be silent
completely. By default, it sets to false. If true, verbose will be
forced to be false}
}
\value{
an object of class "eGSEA", a list with following components:
\itemize{
\item{\code{df_summary}: a data frame of nTerm x 12 containing gene set
enrichment analysis result, where nTerm is the number of
terms/genesets, and the 9 columns are "setID" (i.e. "Term ID"), "nAnno"
(i.e. number in members annotated by a term), "nLead" (i.e. number in
members as leading genes), "peak" (i.e. the rank at peak), "total"
(i.e. the total number of genes analysed), "es" (i.e. enrichment
score), "nes" (i.e. normalised enrichment score; enrichment score but
after being normalised by gene set size), "pvalue" (i.e. nominal p
value), "adjp" (i.e. adjusted p value; p value but after being adjusted
for multiple comparisons), "frac" (nLead/nAnno), "member" (members at
the leading edge), and "member_rank" (members and their ranks)}
\item{\code{leading}: a list of gene sets, each storing leading gene
info (i.e. the named vector with names for gene symbols and elements
for priority rank). Always, gene sets are identified by "setID"}
\item{\code{full}: a list of gene sets, each storing full info on gene
set enrichment analysis result (i.e. a data frame of nGene x 6, where
nGene is the number of genes, and the 6 columns are "GeneID", "Rank"
for priority rank, "Score" for priority score, "RES" for running
enrichment score, "Hits" for gene set hits info with 1 for gene hit, 2
for leading gene hit, 3 for the point defining leading genes, 0 for no
hit). Always, gene sets are identified by "setID"}
\item{\code{cross}: a matrix of nTerm X nTerm, with an on-diagnal cell
for the leading genes observed in an individaul term, and off-diagnal
cell for the overlapped leading genes shared between two terms}
}
}
\description{
\code{oPierGSEA} is supposed to prioritise pathways given prioritised
genes and the ontology in query. It is done via gene set enrichment
analysis (GSEA). It returns an object of class "eGSEA".
}
\note{
none
}
\examples{
\dontrun{
# a) provide the seed nodes/genes with the weight info
## get genes within 500kb away from AS GWAS lead SNPs
seeds.genes <- ImmunoBase$AS$genes_variants
## seeds weighted according to distance away from lead SNPs
data <- 1- seeds.genes/500000

# b) perform priority analysis
pNode <- oPierGenes(data=data, network="PCommonsDN_medium",restart=0.7)

# c) do pathway-level priority using GSEA
eGSEA <- oPierGSEA(pNode=pNode, ontology="DGIdb", nperm=2000,
placeholder=placeholder)
bp <- oGSEAbarplot(eGSEA, top_num="auto", displayBy="nes")
gp <- oGSEAdotplot(eGSEA, top=1)
}
}
\seealso{
\code{\link{oPierGSEA}}
}
