% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oPierManhattan.r
\name{oPierManhattan}
\alias{oPierManhattan}
\title{Function to visualise prioritised genes using manhattan plot}
\usage{
oPierManhattan(
pNode,
color = c("darkred", "steelblue4"),
point.size = 0.2,
top = 10000,
top.label.type = c("text", "box"),
top.label.size = 2,
top.label.col = "black",
top.label.query = NULL,
label.query.only = FALSE,
chromosome.only = TRUE,
y.scale = c("normal", "sqrt", "log"),
y.lab = NULL,
GR.Gene = c("UCSC_knownGene", "UCSC_knownCanonical"),
font.family = "sans",
verbose = TRUE,
placeholder = NULL,
guid = NULL,
...
)
}
\arguments{
\item{pNode}{an object of class "pNode" (or "sTarget" or "dTarget")}

\item{color}{a character vector for colors to alternate chromosome
colorings. If NULL, ggplot2 default colors will be used. If a single
character is provided, it can be "jet" (jet colormap) or "rainbow"
(rainbow colormap, that is, red-yellow-green-cyan-blue-magenta)}

\item{point.size}{the point size}

\item{top}{the number of the top targets to be labelled/highlighted}

\item{top.label.type}{how to label the top targets. It can be "box"
drawing a box around the labels , and "text" for the text only}

\item{top.label.size}{the highlight label size}

\item{top.label.col}{the highlight label color}

\item{top.label.query}{which top genes in query will be labelled. By
default, it sets to NULL meaning all top genes will be displayed. If
labels in query can not be found, then all will be displayed}

\item{label.query.only}{logical to indicate whether only genes in query
will be displayed. By default, it sets to FALSE. It only works when
labels in query are enabled/found}

\item{chromosome.only}{logical to indicate whether only genes from
input data will be displayed. By default, it sets to TRUE}

\item{y.scale}{how to transform the y scale. It can be "normal" for no
transformation, "sqrt" for square root transformation, and "log" for
log-based transformation}

\item{y.lab}{the y labelling. If NULL (by default), it shows the column
of input data}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_knownGene', that is, UCSC known genes (together with genomic
locations) based on human genome assembly hg19. It can be
'UCSC_knownCanonical', that is, UCSC known canonical genes (together
with genomic locations) based on human genome assembly hg19.
Alternatively, the user can specify the customised input. To do so,
first save your RData file (containing an GR object) into your local
computer, and make sure the GR object content names refer to Gene
Symbols. Then, tell "GR.Gene" with your RData file name (with or
without extension), plus specify your file RData path in "placeholder"}

\item{font.family}{the font family for texts}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}

\item{placeholder}{the characters to tell the location of built-in
RData files. See \code{\link{oRDS}} for details}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF
project. See \code{\link{oRDS}} for details}

\item{...}{additional paramters associated with
ggrepel::geom_text_repel}
}
\value{
an object of class "ggplot", appended by an GR object called 'gr'
}
\description{
\code{oPierManhattan} is supposed to visualise prioritised genes using
manhattan plot. Genes with the top priority are highlighed. It returns
an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
mp <- oPierManhattan(pNode, placeholder=placeholder)
mp$gr
## control visuals
mp <- oPierManhattan(pNode, color='ggplot2', top=50,
top.label.col="black", y.scale="sqrt", placeholder=placeholder)
mp
## control labels
# only IL genes will be labelled
ind <- grep('^IL', rownames(pNode$priority))
top.label.query <- rownames(pNode$priority)[ind]
mp <- oPierManhattan(pNode, top.label.query=top.label.query,
placeholder=placeholder)
mp
# only IL genes will be displayed
mp <- oPierManhattan(pNode, top.label.query=top.label.query,
label.query.only=TRUE, placeholder=placeholder)
mp
}
}
\seealso{
\code{\link{oRDS}}, \code{\link{oColormap}}
}
