% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oPierMatrix.r
\name{oPierMatrix}
\alias{oPierMatrix}
\title{Function to extract priority or evidence matrix from a list of pNode objects}
\usage{
oPierMatrix(
list_pNode,
displayBy = c("score", "rank", "weight", "pvalue", "evidence"),
combineBy = c("union", "intersect"),
aggregateBy = c("none", "fishers", "logistic", "Ztransform",
"orderStatistic",
"harmonic", "max", "sum"),
rangeMax = 5,
keep = FALSE,
GR.Gene = c("UCSC_knownGene", "UCSC_knownCanonical"),
verbose = TRUE,
placeholder = NULL,
guid = NULL
)
}
\arguments{
\item{list_pNode}{a list of "pNode" objects or a "pNode" object}

\item{displayBy}{which priority will be extracted. It can be "score"
for priority score/rating (by default), "rank" for priority rank,
"weight" for seed weight, "pvalue" for priority p-value, "evidence" for
the evidence (seed info)}

\item{combineBy}{how to resolve nodes/targets from a list of "pNode"
objects. It can be "intersect" for intersecting nodes (by default),
"union" for unionising nodes}

\item{aggregateBy}{the aggregate method used. It can be either "none"
for no aggregation, or "orderStatistic" for the method based on the
order statistics of p-values, "fishers" for Fisher's method,
"Ztransform" for Z-transform method, "logistic" for the logistic
method. Without loss of generality, the Z-transform method does well in
problems where evidence against the combined null is spread widely
(equal footings) or when the total evidence is weak; Fisher's method
does best in problems where the evidence is concentrated in a
relatively small fraction of the individual tests or when the evidence
is at least moderately strong; the logistic method provides a
compromise between these two. Notably, the aggregate methods 'fishers'
and 'logistic' are preferred here. Also supported are methods summing
up evidence 'sum', taking the maximum of evidence ('max') or
sequentially weighting evidence 'harmonic'}

\item{rangeMax}{the maximum range of the top prioritisation. By
default, it sets to 5}

\item{keep}{logical to indicate whether the input list_pNode is kept.
By default, it sets to true to keep}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_knownGene', that is, UCSC known genes (together with genomic
locations) based on human genome assembly hg19. It can be
'UCSC_knownCanonical', that is, UCSC known canonical genes (together
with genomic locations) based on human genome assembly hg19.
Alternatively, the user can specify the customised input. To do so,
first save your RData file (containing an GR object) into your local
computer, and make sure the GR object content names refer to Gene
Symbols. Then, tell "GR.Gene" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location". Note: you can also load your customised GR object
directly}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{placeholder}{the characters to tell the location of built-in
RData files. See \code{\link{oRDS}} for details}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF
project. See \code{\link{oRDS}} for details}
}
\value{
If displayBy is 'evidence', an object of the class "eTarget", a list
with following components:
\itemize{
\item{\code{evidence}: a data frame of nGene X 6 containing gene
evidence information, where nGene is the number of genes, and the 7
columns are seed info including "Overall" for the total number of
different types of seeds, followed by details on individual type of
seeds (that is, "dGene", "pGene", "fGene", "nGene", "eGene", "cGene")}
\item{\code{metag}: an "igraph" object}
}
Otherwise (if displayBy is not 'evidence'), if aggregateBy is 'none'
(by default), a data frame containing priority matrix, with each
column/predictor for either priority score, or priorty rank or priority
p-value. If aggregateBy is not 'none', an object of the class
"dTarget", a list with following components:
\itemize{
\item{\code{priority}: a data frame of n X 4+7 containing gene priority
(aggregated) information, where n is the number of genes, and the 4
columns are "name" (gene names), "rank" (ranks of the priority scores),
"rating" (the 5-star score/rating), "description" (gene description),
and 7 seed info columns including "seed" (whether or not seed genes),
"nGene" (nearby genes), "cGene" (conformation genes), "eGene" (eQTL
gens), "dGene" (disease genes), "pGene" (phenotype genes), and "fGene"
(function genes)}
\item{\code{predictor}: a data frame containing predictor matrix, with
each column/predictor for either priority score/rating, or priorty rank
or priority p-value}
\item{\code{metag}: an "igraph" object}
\item{\code{list_pNode}: a list of "pNode" objects or NULL}
}
}
\description{
\code{oPierMatrix} is supposed to extract priority or evidence matrix
from a list of pNode objects. Also supported is the aggregation of
priority matrix (similar to the meta-analysis) generating the priority
results; we view this functionality as the discovery mode of the
prioritisation.
}
\note{
none
}
\examples{
\dontrun{
# get predictor matrix for targets
df_score <- oPierMatrix(ls_pNode)
# get evidence for targets
eTarget <- oPierMatrix(ls_pNode, displayBy="evidence")
# get target priority in a discovery mode
dTarget <- oPierMatrix(ls_pNode, displayBy="pvalue",
aggregateBy="fishers")
}
}
\seealso{
\code{\link{oPierMatrix}}
}
