% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSEAextract.r
\name{oSEAextract}
\alias{oSEAextract}
\title{Function to extract enrichment results}
\usage{
oSEAextract(
obj,
sortBy = c("adjp", "pvalue", "zscore", "fc", "nA", "nO", "or", "id",
"name",
"distance")
)
}
\arguments{
\item{obj}{an object of class "eSET" or "eSAD"}

\item{sortBy}{which statistics will be used for sorting and viewing
gene sets (terms). It can be "adjp" for adjusted p value (FDR),
"pvalue" for p value, "zscore" for enrichment z-score, "fc" for
enrichment fold change, "nA" for the number of sets (terms), "nO" for
the number in overlaps, "or" for the odds ratio, "name" for set name,
"id" for set ID and "distance" for set distance}
}
\value{
a tibble with 14 columns:
\itemize{
\item{\code{name}: set name}
\item{\code{id}: set ID}
\item{\code{nA}: number of annotations}
\item{\code{nO}: number of overlaps}
\item{\code{fc}: enrichment fold change}
\item{\code{zscore}: enrichment z-score}
\item{\code{pvalue}: nominal p-value}
\item{\code{adjp}: adjusted p-value (FDR)}
\item{\code{or}: odds ratio}
\item{\code{CIl}: lower bound confidence interval for the odds ratio}
\item{\code{CIu}: upper bound confidence interval for the odds ratio}
\item{\code{distance}: set distance to the root}
\item{\code{namespace}: set namespace}
\item{\code{overlap}: semi-colon separated members for overlaps}
}
}
\description{
\code{oSEAextract} is supposed to extract results of enrichment
analysis.
}
\note{
none
}
\examples{
\dontrun{
oSEAextract(obj)
}
}
\seealso{
\code{\link{oSEAextract}}
}
