% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSEAforest.r
\name{oSEAforest}
\alias{oSEAforest}
\title{Function to visualise enrichment results using a forest plot}
\usage{
oSEAforest(
obj,
top = 10,
adjp.cutoff = 0.05,
colormap = "brewer.Reds",
zlim = NULL,
color.title = expression(-log[10]("FDR")),
shape = 18,
size.range = c(0, 1),
slim = NULL,
size.title = "Overlap",
wrap.width = NULL,
legend.direction = c("auto", "horizontal", "vertical"),
sortBy = c("or", "none")
)
}
\arguments{
\item{obj}{an object of class "eSET" or "eSAD". Alterntively, it can be
a tibble having all these columns (named as
'name','adjp','or','CIl','CIu'); optionally ('group', 'namespace',
'onto', 'nO')}

\item{top}{the number of the top terms (sorted according to OR). For
the "eSET" object, if it is 'auto' (for eSET), only the significant
terms (see below adjp.cutoff) will be displayed}

\item{adjp.cutoff}{FDR cutoff used to declare the significant terms. By
default, it is set to 0.05. Only works when top is 'auto' above}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{https://html-color-codes.info/color-names/index.html}}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted, defaulting to the range of the -log10(adjp)}

\item{color.title}{the legend title for colorbar}

\item{shape}{the point shape}

\item{size.range}{the point size range}

\item{slim}{the minimum and maximum point size values, defaulting to
the range of the nO}

\item{size.title}{the size legend title}

\item{wrap.width}{a positive integer specifying wrap width of name}

\item{legend.direction}{the legend guide direction. It can be
"horizontal", "vertical" and "auto" ("vertical" when having >15 rows;
otherwise "horizontal")}

\item{sortBy}{It can be "or" or "none"}
}
\value{
an object of class "ggplot"
}
\description{
\code{oSEAforest} is supposed to visualise enrichment results using a
forest plot. A point is colored by the significance level, and a
horizontal line for the 95% confidence interval (CI) of odds ratio (OR;
the wider the CI, the less reliable). It returns an object of class
"ggplot".
}
\note{
none
}
\examples{
\dontrun{
obj \%>\% oSEAforest()
obj \%>\% oSEAextract() \%>\% filter(onto=='PSG') \%>\% oSEAforest()
}
}
\seealso{
\code{\link{oSEAforest}}
}
