% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSEAggraph.r
\name{oSEAggraph}
\alias{oSEAggraph}
\title{Function to visualise enrichment results using a ggraph-like lauout}
\usage{
oSEAggraph(
obj,
ig = NULL,
fixed = FALSE,
node.color = c("zscore", "adjp", "or"),
colormap = "grey90-orange-darkred",
zlim = NULL,
node.size = c("adjp", "zscore", "or"),
slim = NULL,
node.size.range = c(0.5, 4),
node.label.size = 2,
leave = TRUE,
ncolumns = NULL,
...
)
}
\arguments{
\item{obj}{an object of class "eset" or "ls_eset". Alterntively, it can
be a data frame having all these columns ('name','adjp','or','zscore';
'group' optionally)}

\item{ig}{an object of class "igraph" with node attribute 'name'. Note:
the node labels would be the node attribute 'name' unless the node
attribute 'label' is explicitely provided. If provided, only those
terms within it will be visualised. By default, it is NULL meaning no
surch restriction}

\item{fixed}{logical to indicate whether all terms in ig will be
visualised. By default, it is TURE; otherwise only overlapped terms
from eset will be visualised}

\item{node.color}{which statistics will be used for node coloring. It
can be "or" for the odds ratio, "adjp" for adjusted p value (FDR) and
"zscore" for enrichment z-score}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap),
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta), and "ggplot2" (emulating ggplot2
default color palette). Alternatively, any hyphen-separated HTML color
names, e.g. "lightyellow-orange" (by default), "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{https://html-color-codes.info/color-names/index.html}}

\item{zlim}{the minimum and maximum values for which colors should be
plotted}

\item{node.size}{which statistics will be used for node size. It can be
"or" for the odds ratio, "adjp" for adjusted p value (FDR) and "zscore"
for enrichment z-score}

\item{slim}{the minimum and maximum values for which sizes should be
plotted}

\item{node.size.range}{the range of actual node size}

\item{node.label.size}{the text size of the node labelings. By default,
it is 2. If 0, all labellings will be disabled}

\item{leave}{the logic specifying whether or not only leaves
(nodes/labellings) shown. This can be disenabled if the layout does not
support tips}

\item{ncolumns}{an integer specifying the number of columns for
facet_wrap. By defaul, it is NULL (decided on according to the number
of groups that will be visualised)}

\item{...}{additional graphic parameters used in oGGraph}
}
\value{
a ggplot2 object appended with 'ig', 'data' which should contain
columns 'x','y', 'name' (the same as V(ig)$name), 'label' (if not given
in ig, a 'name' varient), 'data_enrichment' (enrichment results), and
'gp_template' with labelling if multiple groups (together with no
labelling for the colored plots).
}
\description{
\code{oSEAggraph} is supposed to visualise enrichment results using a
ggraph-like lauout.
}
\note{
none
}
\examples{
\dontrun{
# circular visualisation of enriched AA pathways
gp <- oSEAggraph(eset, ig)

###############################
# advanced use: multiple groups

# circular visualisation of enriched AA pathways
gp <- oSEAggraph(ls_eset, ig)
gp
gp$gp_template
}
}
\seealso{
\code{\link{oGGraph}}
}
