% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSEAupset.r
\name{oSEAupset}
\alias{oSEAupset}
\title{Function to visualise enrichment results using a upset plot}
\usage{
oSEAupset(
obj,
top = 10,
adjp.cutoff = 0.05,
color = "steelblue4",
shape = 18,
size.range = c(2, 5),
slim = NULL,
size.title = "Overlap",
wrap.width = NULL,
label.height.unit = 5.5,
member.levels = c("auto", "num", "customised"),
levels.customised = NULL,
sortBy = c("pvalue", "name")
)
}
\arguments{
\item{obj}{an object of class "eSET". Alterntively, it can be a tibble
having all these columns (named as 'name','adjp','nO','overlap')}

\item{top}{the number of the top terms (sorted according to adjp). For
the "eSET" object, if it is 'auto' (for eSET), only the significant
terms (see below adjp.cutoff) will be displayed}

\item{adjp.cutoff}{FDR cutoff used to declare the significant terms. By
default, it is set to 0.05. Only works when top is 'auto' above}

\item{color}{the point color}

\item{shape}{the point shape}

\item{size.range}{the point size range}

\item{slim}{the minimum and maximum point size values, defaulting to
the range of the nO}

\item{size.title}{the size legend title}

\item{wrap.width}{a positive integer specifying wrap width of name}

\item{label.height.unit}{the unit specifying the per-row height of the
combination matrix. By default, it is NULL. If specified (such as 8),
it will be used to decide 'combmatrix.label.height' and
'combmatrix.label.text'}

\item{member.levels}{how to define the levels of members in the
combination matri. It can be 'num' (the number of non-zeros for a
member) or 'customised' (see 'levels.customised' below) or 'auto'
(alphabeta order; by default)}

\item{levels.customised}{the customised levels of members in the
combination matrix. Only works when member.levels is 'customised'
above}

\item{sortBy}{how to sort bars/combinations. It can be 'pvalue' or
'name'}
}
\value{
an object of class "ggplot"
}
\description{
\code{oSEAupset} is supposed to visualise enrichment results using a
upset plot. The top is the kite plot, and visualised below is the
combination matrix for overlapped genes. It returns an object of class
"ggplot".
}
\note{
none
}
\examples{
\dontrun{
obj \%>\% oSEAupset()
obj \%>\% oSEAextract() \%>\% filter(onto=='PSG') \%>\% oSEAupset()
}
}
\seealso{
\code{\link{oSEAupset}}
}
