% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSNP2cGenes.r
\name{oSNP2cGenes}
\alias{oSNP2cGenes}
\title{Function to define HiC genes given a list of SNPs}
\usage{
oSNP2cGenes(
data,
entity = c("SNP", "chr:start-end", "data.frame", "bed", "GRanges"),
include.RGB = NA,
GR.SNP = c("dbSNP_GWAS", "dbSNP_Common"),
GR.Gene = c("UCSC_knownGene", "UCSC_knownCanonical"),
cdf.function = c("empirical", "exponential"),
verbose = TRUE,
placeholder = NULL,
guid = NULL
)
}
\arguments{
\item{data}{an input vector containing SNPs. SNPs should be provided as
dbSNP ID (ie starting with rs) or in the format of 'chrN:xxx', where N
is either 1-22 or X, xxx is number; for example, 'chr16:28525386'.
Alternatively, it can be other formats/entities (see the next parameter
'entity')}

\item{entity}{the data entity. By default, it is "SNP". For general
use, it can also be one of "chr:start-end", "data.frame", "bed" or
"GRanges"}

\item{include.RGB}{genes linked to input SNPs are also included. By
default, it is 'NA' to disable this option. Otherwise, those genes
linked to SNPs will be included according to Promoter Capture HiC
(PCHiC) datasets. Pre-built HiC datasets are detailed in
\code{\link{oDefineRGB}}}

\item{GR.SNP}{the genomic regions of SNPs. By default, it is
'dbSNP_GWAS', that is, SNPs from dbSNP (version 146) restricted to GWAS
SNPs and their LD SNPs (hg19). It can be 'dbSNP_Common', that is,
Common SNPs from dbSNP (version 146) plus GWAS SNPs and their LD SNPs
(hg19). Alternatively, the user can specify the customised input. To do
so, first save your RData file (containing an GR object) into your
local computer, and make sure the GR object content names refer to
dbSNP IDs. Then, tell "GR.SNP" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location". Note: you can also load your customised GR object
directly}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_knownGene', that is, UCSC known genes (together with genomic
locations) based on human genome assembly hg19. It can be
'UCSC_knownCanonical', that is, UCSC known canonical genes (together
with genomic locations) based on human genome assembly hg19.
Alternatively, the user can specify the customised input. To do so,
first save your RData file (containing an GR object) into your local
computer, and make sure the GR object content names refer to Gene
Symbols. Then, tell "GR.Gene" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location". Note: you can also load your customised GR object
directly}

\item{cdf.function}{a character specifying a Cumulative Distribution
Function (cdf). It can be one of 'exponential' based on exponential
cdf, 'empirical' for empirical cdf}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{placeholder}{the characters to tell the location of built-in
RData files. See \code{\link{oRDS}} for details}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF
project. See \code{\link{oRDS}} for details}
}
\value{
a data frame with following columns:
\itemize{
\item{\code{Gene}: SNP-interacting genes caputured by HiC}
\item{\code{SNP}: SNPs}
\item{\code{Sig}: the interaction score (the higher stronger)}
\item{\code{Weight}: the HiC weight}
}
}
\description{
\code{oSNP2cGenes} is supposed to define HiC genes given a list of
SNPs. The HiC weight is calcualted as Cumulative Distribution Function
of HiC interaction scores.
}
\note{
none
}
\examples{
\dontrun{
# a) provide the SNPs with the significance info
data <- names(ImmunoBase$AS$variants)

# b) define HiC genes
df_cGenes <- oSNP2cGenes(data,
include.RGB="PCHiC_PMID27863249_Monocytes", placeholder=placeholder)
}
}
\seealso{
\code{\link{oSNP2cGenes}}
}
