% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSNP2eGenes.r
\name{oSNP2eGenes}
\alias{oSNP2eGenes}
\title{Function to define eQTL genes given a list of SNPs or a customised eQTL mapping data}
\usage{
oSNP2eGenes(
data,
include.QTL = NA,
QTL.customised = NULL,
GR.Gene = c("UCSC_knownGene", "UCSC_knownCanonical"),
cdf.function = c("empirical", "exponential"),
plot = FALSE,
verbose = TRUE,
placeholder = NULL,
guid = NULL
)
}
\arguments{
\item{data}{an input vector containing SNPs. SNPs should be provided as
dbSNP ID (ie starting with rs). Alternatively, they can be in the
format of 'chrN:xxx', where N is either 1-22 or X, xxx is number; for
example, 'chr16:28525386'}

\item{include.QTL}{the eQTL supported currently. By default, it is 'NA'
to disable this option. Pre-built eQTL datasets are detailed in
\code{\link{oDefineQTL}}}

\item{QTL.customised}{a user-input matrix or data frame with 4 columns:
1st column for SNPs/eQTLs, 2nd column for Genes, 3rd for eQTL mapping
significance level (p-values or FDR), and 4th for contexts (required
even though only one context is input). Alternatively, it can be a file
containing these 4 columns. It is designed to allow the user analysing
their eQTL data. This customisation (if provided) will populate
built-in eQTL data}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_knownGene', that is, UCSC known genes (together with genomic
locations) based on human genome assembly hg19. It can be
'UCSC_knownCanonical', that is, UCSC known canonical genes (together
with genomic locations) based on human genome assembly hg19.
Alternatively, the user can specify the customised input. To do so,
first save your RData file (containing an GR object) into your local
computer, and make sure the GR object content names refer to Gene
Symbols. Then, tell "GR.Gene" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location". Note: you can also load your customised GR object
directly}

\item{cdf.function}{a character specifying a Cumulative Distribution
Function (cdf). It can be one of 'exponential' based on exponential
cdf, 'empirical' for empirical cdf}

\item{plot}{logical to indicate whether the histogram plot (plus
density or CDF plot) should be drawn. By default, it sets to false for
no plotting}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{placeholder}{the characters to tell the location of built-in
RData files. See \code{\link{oRDS}} for details}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF
project. See \code{\link{oRDS}} for details}
}
\value{
a data frame with following columns:
\itemize{
\item{\code{Gene}: eQTL-containing genes}
\item{\code{SNP}: eQTLs}
\item{\code{Sig}: the eQTL mapping significant level (the
best/minimum)}
\item{\code{Weight}: the eQTL weight}
}
}
\description{
\code{oSNP2eGenes} is supposed to define eQTL genes given a list of
SNPs or a customised eQTL mapping data. The eQTL weight is calcualted
as Cumulative Distribution Function of negative log-transformed
eQTL-reported signficance level.
}
\note{
none
}
\examples{
\dontrun{
# a) provide the SNPs with the significance info
gr <- ImmunoBase$AS$variants
AS <- as.data.frame(GenomicRanges::mcols(gr)[, c('Variant','Pvalue')])

# b) define eQTL genes
df_eGenes <- oSNP2eGenes(data=AS[,1], include.QTL="JKscience_TS2A")
}
}
\seealso{
\code{\link{oSNP2eGenes}}
}
