% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSymbol2GeneID.r
\name{oSymbol2GeneID}
\alias{oSymbol2GeneID}
\title{Function to convert gene symbols to entrez geneid}
\usage{
oSymbol2GeneID(
data,
org = c("human", "mouse"),
check.symbol.identity = FALSE,
details = FALSE,
verbose = TRUE,
placeholder = NULL,
guid = NULL
)
}
\arguments{
\item{data}{an input vector containing gene symbols}

\item{org}{a character specifying an organism. Currently supported
organisms are 'human' and 'mouse'. It can be an object 'EG'}

\item{check.symbol.identity}{logical to indicate whether to match the
input data via Synonyms for those unmatchable by official gene symbols.
By default, it sets to false}

\item{details}{logical to indicate whether to result in a data frame
(in great details). By default, it sets to false}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}

\item{placeholder}{the characters to tell the location of built-in
RData files. See \code{\link{oRDS}} for details}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF
project. See \code{\link{oRDS}} for details}
}
\value{
a vector containing entrez geneid with 'NA' for the unmatched if
(details set to false); otherwise, a data frame is returned
}
\description{
\code{oSymbol2GeneID} is supposed to convert gene symbols to entrez
geneid.
}
\note{
If a symbol mapped many times, the one assiged as the "protein-coding"
type of gene is preferred.
}
\examples{
\dontrun{

# a) provide the input Genes of interest (eg 100 randomly chosen human genes)
## load human genes
org.Hs.eg <- oRDS('org.Hs.eg', placeholder=placeholder)
Symbol <- sample(org.Hs.eg$info$Symbol, 100)
Symbol

# b) convert into GeneID
GeneID <- oSymbol2GeneID(Symbol)

# c) convert into a data frame
df <- oSymbol2GeneID(Symbol, details=TRUE)


# advanced use
df <- oSymbol2GeneID(Symbol, org=org.Hs.eg, details=TRUE)
}
}
\seealso{
\code{\link{oSymbol2GeneID}}
}
