% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oTIG.r
\name{oTIG}
\alias{oTIG}
\title{Function to infer relations between terms based on shared members}
\usage{
oTIG(
data,
method = c("hybrid", "jaccard", "empirical"),
empirical.cutoff = 0.5
)
}
\arguments{
\item{data}{a tibble with two columns 'name' and 'members' (each is a
vector containing members separated by ', '). Alternatively a data
frame (matrix) with rows for members (row names) and columns for terms
(column names)}

\item{method}{how to infer relations between two terms. It can be
'jaccard' (minimum spanning tree minimising the jaccard distance),
'empirical' (the number of the shared members is no less than
'empirical.cutoff' of two-end terms), and 'hybrid' (the union of
results from 'jaccard' and 'empirical')}

\item{empirical.cutoff}{the empirical cutoff used to declare the
relation between two terms (the number of the shared members is no less
than 'empirical.cutoff' of two-end terms). By default, it is set to
0.5. Only works when the method is 'empirical' or 'hybrid' above}
}
\value{
an object of class "igraph" with node attributes ('name', 'members',
'n_members', 'xcoord' and 'ycoord') and edge attributes ('shared',
'n_shared' and 'weight')
}
\description{
\code{oTIG} is supposed to infer relations between terms based on
shared members. It returns an object of class "igraph".
}
\note{
none
}
\examples{
set.seed(825)
mat <- ifelse(matrix(rnorm(1000),200,5)>0, 1, 0)
rownames(mat) <- paste0('R', seq(nrow(mat)))
colnames(mat) <- paste0('C', seq(ncol(mat)))

\dontrun{
tig <- oTIG(mat, method='hybrid')
gp <- oGGnetwork(tig, node.label='name', node.label.size=2.5,
node.label.color='darkblue', node.label.force=1, node.xcoord='xcoord',
node.ycoord='ycoord', colormap="cyan4-cyan4", edge.size='weight',
node.shape=18, node.size='n_members', node.size.title="Num of \ngenes",
node.size.range=c(2,5), edge.color='cyan4', edge.color.alpha=0.3,
edge.curve=0,edge.arrow.gap=0.01)
}
}
\seealso{
\code{\link{oTIG}}
}
