% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oUpsetAdv.r
\name{oUpsetAdv}
\alias{oUpsetAdv}
\title{Function to visualise enrichment results using a upset plot}
\usage{
oUpsetAdv(
data,
sortBy = c("auto", "value", "degree"),
color = "cyan4",
shape = 18,
size = 2,
label.height.unit = NULL,
member.levels = c("auto", "num", "customised"),
levels.customised = NULL
)
}
\arguments{
\item{data}{a data frame. It contains all these columns including
'value' and 'member' (separated by ',' or ', ')}

\item{sortBy}{how to sort bars/combinations. It can be 'degree' (number
of non-zeros for a combination) or 'value' (value for a combination) or
'auto' (the same order as provided; by default)}

\item{color}{the point color}

\item{shape}{the point shape}

\item{size}{the point size}

\item{label.height.unit}{the unit specifying the per-row height of the
combination matrix. By default, it is NULL. If specified (such as 8),
it will be used to decide 'combmatrix.label.height' and
'combmatrix.label.text'}

\item{member.levels}{how to define the levels of members in the
combination matri. It can be 'num' (the number of non-zeros for a
member) or 'customised' (see 'levels.customised' below) or 'auto'
(alphabeta order; by default)}

\item{levels.customised}{the customised levels of memebers in the
combination matri. Only works when member.levels is 'customised' above}
}
\value{
an object of class "ggplot"
}
\description{
\code{oUpsetAdv} is supposed to visualise enrichment results using a
upset plot. The top is the kite plot, and visualised below is the
combination matrix for overlapped genes. It returns an object of class
"ggplot".
}
\note{
none
}
\examples{
\dontrun{
obj \%>\% transmute(value=nO, member=overlap) -> data
#data \%>\% arrange(-value) \%>\% mutate(member=fct_inorder(member)) -> data
data \%>\% oUpsetAdv() -> gp
gp + geom_point(aes(fill=as.factor(i)), shape=22, size=2,
color="transparent") + geom_line(aes(group=i,color=as.factor(i)))

# customised levels for members
data \%>\% separate_rows(member,sep=',') \%>\% arrange(member) \%>\%
pull(member) \%>\% unique() -> levels
data \%>\%
oUpsetAdv(member.levels='customised',levels.customised=levels)
}
}
\seealso{
\code{\link{oUpsetAdv}}
}
