% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-PIG.r
\name{pNode}
\alias{pNode}
\alias{print.pNode}
\title{Definition for S3 class \code{pNode}}
\usage{
pNode(priority, g)

\method{print}{pNode}(x, ...)
}
\arguments{
\item{priority}{a data frame}

\item{g}{an 'igraph' object}

\item{x}{an object of class \code{pNode}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{pNode}
}
\description{
\code{pNode} has 7 components: priority, g, SNP, Gene2SNP, nGenes,
eGenes and cGenes.
}
\examples{
\dontrun{
pNode(evidence, metag)
}
}
\keyword{S3}
\keyword{classes}
